requirements.txt
beautifulsoup4==4.9.3
bidict==0.23.1
cachelib==0.13.0
cachetools==4.2.4
certifi==2025.4.26
charset-normalizer==2.0.12
click==8.1.8
colorama==0.4.6
feedparser==6.0.11
Flask==2.0.1
Flask-Caching==2.3.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
praw==7.7.1
google-api-python-client==2.104.0
google-api-core==1.34.1
# google-api-python-client==2.0.2
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
h11==0.16.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==3.20.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==0.19.0
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.26.0
rsa==4.9.1
setuptools==80.3.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.7
# SQLAlchemy==2.0.40
SQLAlchemy==1.4.48
typing_extensions==4.13.2
uritemplate==3.0.1
urllib3==1.26.20
webencodings==0.5.1
Werkzeug==2.3.0
wsproto==1.2.0
